:heavy_plus_sign: chore(utils): add ne_110m_admin_0_countries.zip for map creation
Added `ne_110m_admin_0_countries.zip` to the project, which contains all the
necessary shapefiles for creating country maps. This was required due to
recent changes in GeoPandas, where the built-in dataset is no longer directly
accessible. The shapefiles enable accurate visualization and mapping of country
boundaries for the happiness index heatmap.